+++ tests/records.v
@@ -1,3 +1,4 @@
+From HB Require Import structures.
 From mathcomp Require Import all_ssreflect.
 From deriving Require Import deriving.
 
@@ -6,9 +7,11 @@ Scheme foo_rect := Induction for foo Sort Type.
 
 Definition foo_indDef := [indDef for foo_rect].
 Canonical foo_indType := IndType foo foo_indDef.
-Definition foo_eqMixin := [derive eqMixin for foo].
-Canonical foo_eqType := EqType foo foo_eqMixin.
-Definition foo_choiceMixin := [derive choiceMixin for foo].
-Canonical foo_choiceType := Eval hnf in ChoiceType foo foo_choiceMixin.
-Definition foo_countMixin := [derive countMixin for foo].
-Canonical foo_countType := Eval hnf in CountType foo foo_countMixin.
+Definition foo_hasDecEq := [derive hasDecEq for foo].
+HB.instance Definition _ := foo_hasDecEq.
+Definition foo_hasChoice := [derive hasChoice for foo].
+HB.instance Definition _ := foo_hasChoice.
+Definition foo_isCountable := [derive isCountable for foo].
+HB.instance Definition _ := foo_isCountable.
+Definition foo_isOrder := [derive isOrder for foo].
+HB.instance Definition _ := foo_isOrder.